common/requirements.txt
openai
langchain
langchain-openai
langchain-community
langchain-experimental
langgraph
langserve[all]
langchain-cli
botbuilder-integration-aiohttp
tiktoken
pypdf
sqlalchemy
pyodbc
tabulate
azure-cosmos
streamlit
python-dotenv
azure-ai-formrecognizer
azure-storage-blob
beautifulsoup4
pydantic
azure-cognitiveservices-speech
fastapi
sse_starlette
sseclient-py



